+++ src/QRegexSearch.h
@@ -89,6 +89,7 @@ class QRegexSearch : public RegexSearchBase
 {
 public:
     QRegexSearch();
+    ~QRegexSearch() override;
 
     Sci::Position FindText(Document *doc, Sci::Position minPos, Sci::Position maxPos, const char *s, bool caseSensitive, bool word, bool wordStart, Scintilla::FindOption flags, Sci::Position *length) override;
     const char *SubstituteByPosition(Document *doc, const char *text, Sci::Position *length) override;